Refactor / state

+++ internal/manifest/manifest_test.go
@@ -1,10 +1,11 @@
-package manifest
+package manifest_test
 
 import (
 	"os"
 	"path"
 	"testing"
 
+	"github.com/kyma-project/istio/operator/internal/manifest"
 	"github.com/kyma-project/istio/operator/internal/tests"
 	"github.com/onsi/ginkgo/v2/types"
 
@@ -28,7 +29,7 @@ var _ = ReportAfterSuite("custom reporter", func(report types.Report) {
 	tests.GenerateGinkgoJunitReport("manifest-suite", report)
 })
 
-var _ = Describe("Manifest merge", func() {
+var _ = Describe("Merge", func() {
 	numTrustedProxies := 4
 	istioCR := &v1alpha2.Istio{ObjectMeta: metav1.ObjectMeta{
 		Name:      "istio-test",
@@ -40,72 +41,41 @@ var _ = Describe("Manifest merge", func() {
 			},
 		},
 	}
-	workingDir := "test"
 
-	It("should return error when provided invalid cluster size", func() {
+	DescribeTable("Merge for differnt cluster sizes", func(clusterSize clusterconfig.ClusterSize, shouldError bool, igwMinReplicas int) {
 		// given
-		sut := NewDefaultIstioMerger()
+		sut := manifest.NewDefaultIstioMerger()
 
 		// when
-		mergedIstioOperatorPath, err := sut.Merge(9, istioCR, clusterconfig.ClusterConfiguration{})
+		mergedIstioOperatorPath, err := sut.Merge(clusterSize, istioCR, clusterconfig.ClusterConfiguration{})
 
 		// then
-		Expect(err).Should(HaveOccurred())
-		Expect(mergedIstioOperatorPath).To(BeEmpty())
-	})
+		if shouldError {
+			Expect(err).Should(HaveOccurred())
+			Expect(err.Error()).To(Equal("unsupported cluster size"))
+			Expect(mergedIstioOperatorPath).To(BeEmpty())
+		} else {
+			Expect(err).ShouldNot(HaveOccurred())
+			Expect(mergedIstioOperatorPath).To(Equal(path.Join("/tmp", manifest.MergedIstioOperatorFile)))
 
-	It("should return merged configuration, when there is a Istio CR with valid configuration and using production manifest", func() {
-		// given
-		sut := NewDefaultIstioMerger()
+			iop := readIOP(mergedIstioOperatorPath)
 
-		// when
-		mergedIstioOperatorPath, err := sut.Merge(clusterconfig.Production, istioCR, clusterconfig.ClusterConfiguration{})
+			numTrustedProxies := iop.Spec.MeshConfig.Fields["defaultConfig"].
+				GetStructValue().Fields["gatewayTopology"].GetStructValue().Fields["numTrustedProxies"].GetNumberValue()
+			Expect(numTrustedProxies).To(Equal(float64(numTrustedProxies)))
 
-		// then
-		Expect(err).Should(Not(HaveOccurred()))
-		Expect(mergedIstioOperatorPath).To(Not(BeEmpty()))
-	})
+			Expect(iop.Spec.Components.IngressGateways[0].K8S.HpaSpec.MinReplicas).To(Equal(int32(igwMinReplicas)))
 
-	It("should return error when provided misconfigured default Istio Operator", func() {
-		// given
-		wrongOperator, err := os.ReadFile("test/wrong-operator.yaml")
-		Expect(err).Should(Not(HaveOccurred()))
+			err = os.Remove(mergedIstioOperatorPath)
+			Expect(err).ShouldNot(HaveOccurred())
+		}
+	},
+		Entry("should return error when provided Unknown cluster size", clusterconfig.UnknownSize, true, 0),
+		Entry("should return merged configuration for Evaluation cluster size", clusterconfig.Evaluation, false, 1),
+		Entry("should return merged configuration for Production cluster size", clusterconfig.Production, false, 3),
+	)
 
-		sut := IstioMerger{workingDir, ManifestGetterMock{wrongOperator}}
-
-		// when
-		mergedIstioOperatorPath, err := sut.Merge(clusterconfig.Production, istioCR, clusterconfig.ClusterConfiguration{})
-
-		// then
-		Expect(err).Should(HaveOccurred())
-		Expect(mergedIstioOperatorPath).To(BeEmpty())
-	})
-
-	It("should return merged configuration, when there is a Istio CR with valid configuration and a correct Istio Operator manifest", func() {
-		// given
-		goodOperator, err := os.ReadFile("test/test-operator.yaml")
-		Expect(err).Should(Not(HaveOccurred()))
-
-		sut := IstioMerger{workingDir, ManifestGetterMock{goodOperator}}
-
-		// when
-		mergedIstioOperatorPath, err := sut.Merge(clusterconfig.Production, istioCR, clusterconfig.ClusterConfiguration{})
-
-		// then
-		Expect(err).ShouldNot(HaveOccurred())
-		Expect(mergedIstioOperatorPath).To(Equal(path.Join(workingDir, mergedIstioOperatorFile)))
-
-		iop := readIOP(mergedIstioOperatorPath)
-
-		numTrustedProxies := iop.Spec.MeshConfig.Fields["defaultConfig"].
-			GetStructValue().Fields["gatewayTopology"].GetStructValue().Fields["numTrustedProxies"].GetNumberValue()
-
-		Expect(numTrustedProxies).To(Equal(float64(4)))
-		err = os.Remove(mergedIstioOperatorPath)
-		Expect(err).ShouldNot(HaveOccurred())
-	})
-
-	It("should return merged configuration with overrides when provided", func() {
+	It("should return merged configuration when overrides are provided", func() {
 		// given
 		newCniBinDirPath := "overriden/path"
 
@@ -124,17 +94,14 @@ var _ = Describe("Manifest merge", func() {
 			},
 		}
 
-		goodOperator, err := os.ReadFile("test/test-operator.yaml")
-		Expect(err).Should(Not(HaveOccurred()))
-
-		sut := IstioMerger{workingDir, ManifestGetterMock{goodOperator}}
+		sut := manifest.NewDefaultIstioMerger()
 
 		// when
 		mergedIstioOperatorPath, err := sut.Merge(clusterconfig.Production, istioCR, clusterConfig)
 
 		// then
 		Expect(err).ShouldNot(HaveOccurred())
-		Expect(mergedIstioOperatorPath).To(Equal(path.Join(workingDir, mergedIstioOperatorFile)))
+		Expect(mergedIstioOperatorPath).To(Equal(path.Join("/tmp", manifest.MergedIstioOperatorFile)))
 
 		iop := readIOP(mergedIstioOperatorPath)
 
@@ -156,6 +123,28 @@ var _ = Describe("Manifest merge", func() {
 	})
 })
 
+var _ = Describe("GetIstioVersion", func() {
+	It("should return Istio version and verify production and evaluation manifests have same hub and tag", func() {
+		// given
+		merger := manifest.NewDefaultIstioMerger()
+
+		// when
+		imageVersion, err := merger.GetIstioImageVersion()
+		Expect(err).Should(Not(HaveOccurred()))
+
+		ioProduction := readIOP("../../internal/istiooperator/istio-operator.yaml")
+		Expect(err).Should(Not(HaveOccurred()))
+
+		ioEvaluation := readIOP("../../internal/istiooperator/istio-operator-light.yaml")
+		Expect(err).Should(Not(HaveOccurred()))
+
+		// then
+		Expect(imageVersion.Tag()).To(Equal(ioProduction.Spec.Tag.GetStringValue()))
+		Expect(ioProduction.Spec.Hub).To(Equal(ioEvaluation.Spec.Hub))
+		Expect(ioProduction.Spec.Tag.GetStringValue()).To(Equal(ioEvaluation.Spec.Tag.GetStringValue()))
+	})
+})
+
 func readIOP(iopv1alpha1FilePath string) iopv1alpha1.IstioOperator {
 	iop := iopv1alpha1.IstioOperator{}
 
@@ -166,12 +155,4 @@ func readIOP(iopv1alpha1FilePath string) iopv1alpha1.IstioOperator {
 	Expect(err).ShouldNot(HaveOccurred())
 
 	return iop
-}
-
-type ManifestGetterMock struct {
-	manifest []byte
-}
-
-func (m ManifestGetterMock) GetBytes(_ clusterconfig.ClusterSize) ([]byte, error) {
-	return m.manifest, nil
 }
